backend/requirements.txt
# backend/requirements.txt - Versão corrigida
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
Django==5.2.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
gunicorn==21.2.0
idna==3.7
psycopg2-binary==2.9.9
pycparser==2.22
pyjwt==2.8.0
python3-openid==3.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1

# ✅ PDF e OCR - versões específicas e compatíveis
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.0.0

# ✅ Adicionar dependências que podem estar faltando
python-dateutil==2.8.2
regex==2023.12.25